refactor(editor): 🚚 move some files of editor
