--- packages/editor/src/index.ts
@@ -1,6 +1,0 @@
-import './styles/index.css'
-import './styles/prosemirror.css'
-
-export { default as Editor } from './ui'
-export { TiptapExtensions } from './ui/extensions'
-export { TiptapEditorProps } from './ui/props'
