+++ packages/editor/src/props.ts
