+++ packages/editor/tsup.config.ts
@@ -1,7 +1,7 @@
 import { Options, defineConfig } from 'tsup'
 
 export default defineConfig((options: Options) => ({
-  entry: ['src/index.ts'],
+  entry: ['src/index.tsx'],
   format: ['cjs', 'esm'],
   banner: {
     js: "'use client'",
